combined unit tests for labeling and relabeling into one file
